feat: Refactor FilmsModule and UsersModule imports and exports
