+++ src/films/films.service.ts
@@ -45,7 +45,10 @@ export class FilmsService {
     delete film.releaseDate;
     delete film.finishDate;
 
-    return film;
+    return {
+      film,
+      userReviews: [],
+    };
   }
 
   async updateFilm(id: string, dto: UpdateFilmDto) {
@@ -196,5 +199,111 @@ export class FilmsService {
     }
 
     return list;
+  }
+
+  async addFavoriteFilm(userId: string, filmId: string) {
+    // Check if the film exists
+    const film = await this.findById(filmId);
+    if (!film) {
+      throw new NotFoundException('Film not found');
+    }
+
+    // Check if user already added the film to favorites
+    const existingFavorite = await this.prisma.userFilmFavorite.findFirst({
+      where: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+    if (existingFavorite) {
+      throw new BadRequestException('Film already added to favorites');
+    }
+
+    // Add the film to the user's favorite films
+    await this.prisma.userFilmFavorite.create({
+      data: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+  }
+
+  async addWatchlistFilm(userId: string, filmId: string) {
+    // Check if the film exists
+    const film = await this.findById(filmId);
+    if (!film) {
+      throw new NotFoundException('Film not found');
+    }
+
+    // Check if user already added the film to watchlist
+    const existingWatchlist = await this.prisma.userFilmWatchlist.findFirst({
+      where: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+    if (existingWatchlist) {
+      throw new BadRequestException('Film already added to watchlist');
+    }
+
+    // Add the film to the user's watchlist
+    await this.prisma.userFilmWatchlist.create({
+      data: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+  }
+
+  async removeFavoriteFilm(userId: string, filmId: string) {
+    // Check if the film exists
+    const film = await this.findById(filmId);
+    if (!film) {
+      throw new NotFoundException('Film not found');
+    }
+
+    // Check if the user added the film to favorites
+    const existingFavorite = await this.prisma.userFilmFavorite.findFirst({
+      where: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+    if (!existingFavorite) {
+      throw new BadRequestException('Film not added to favorites');
+    }
+
+    // Remove the film from the user's favorite films
+    await this.prisma.userFilmFavorite.delete({
+      where: {
+        id: existingFavorite.id,
+      },
+    });
+  }
+
+  async removeWatchlistFilm(userId: string, filmId: string) {
+    // Check if the film exists
+    const film = await this.findById(filmId);
+    if (!film) {
+      throw new NotFoundException('Film not found');
+    }
+
+    // Check if the user added the film to watchlist
+    const existingWatchlist = await this.prisma.userFilmWatchlist.findFirst({
+      where: {
+        user_id: userId,
+        film_id: filmId,
+      },
+    });
+    if (!existingWatchlist) {
+      throw new BadRequestException('Film not added to watchlist');
+    }
+
+    // Remove the film from the user's watchlist
+    await this.prisma.userFilmWatchlist.delete({
+      where: {
+        id: existingWatchlist.id,
+      },
+    });
   }
 }
